pyproject.toml
[tool.poetry]
name = "market"
version = "0.1.0"
description = ""
authors = ["Vip-Artyom <snowboard-team@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.2"
pytest = "^7.3.1"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[tool.poetry.group.dev.dependencies]
ansible-vault-win = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
